Refactors top-level module definitions.
In addition to allowing feature selection, this change will help avoid some
potential (possible, not necessarily likely) import cycles.
